fix: removed unused useState, moved calculate percentage to a helpers file and changed all start to flex start in scss files
