+++ src/shared/components/Cuestionario.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useState } from "react";
 import Checkbox from "./reusable/CustomCheckbox/Checkbox";
 import Resultados from "./Resultados";
+import { BooleanObject } from "shared/types/mainPage.types";
+import { calculatePercentage } from "shared/helpers/mainPage.helpers";
 
 const Cuestionario = () => {
-  const [checkboxStates, setCheckboxStates] = useState<{
-    [key: string]: boolean;
-  }>({
+  const [checkboxStates, setCheckboxStates] = useState<BooleanObject>({
     tr1: false,
     tr2: false,
     tr3: false,
@@ -13,35 +13,7 @@ const Cuestionario = () => {
     tr5: false,
     tr6: false,
   });
-
-  // Calculate the percentage based on checkboxStates
-  const calculatePercentage = () => {
-    const trueValuesCount = Object.values(checkboxStates).reduce(
-      (count, value) => (value ? count + 1 : count),
-      0
-    );
-
-    if (trueValuesCount === 0) {
-      return 1;
-    } else if (trueValuesCount === 1) {
-      return 11;
-    } else if (trueValuesCount === 2) {
-      return 22;
-    } else if (trueValuesCount === 3) {
-      return 34;
-    } else if (trueValuesCount === 4) {
-      return 51;
-    } else if (trueValuesCount === 5) {
-      return 55;
-    } else if (trueValuesCount === 6) {
-      return 100;
-    } else {
-      // Handle other cases if needed
-      return 1;
-    }
-  };
-
-  const [percentage, setPercentage] = useState<number>(calculatePercentage());
+  const [percentage, setPercentage] = useState<number>(1);
 
   const handleCheckboxChange = (
     checkboxId: string,
@@ -60,7 +32,7 @@ const Cuestionario = () => {
 
   // Update the percentage whenever checkboxStates changes
   useEffect(() => {
-    const calculatedPercentage = calculatePercentage();
+    const calculatedPercentage = calculatePercentage(checkboxStates);
     setPercentage(calculatedPercentage);
   }, [checkboxStates]);
