+++ src/shared/components/mainPage.styles.scss
@@ -24,7 +24,7 @@
 
 .nav-links-div {
   display: flex;
-  justify-content: start;
+  justify-content: flex-start;
 }
 
 .nav-link {
@@ -111,7 +111,7 @@
     font-weight: bold;
   }
   ul {
-    text-align: start;
+    text-align: flex-start;
     list-style: disc;
     padding-left: 20px;
   }
