+++ src/shared/components/reusable/CustomCheckbox/Checkbox.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import "./Checkbox.styles.scss";
 
 interface CheckboxProps {
