+++ src/shared/types/mainPage.types.ts
@@ -1,0 +1,3 @@
+export type BooleanObject = {
+  [key: string]: boolean;
+};
